requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
Flask-Mail
Flask-Limiter
WTForms
email-validator
Werkzeug
python-dotenv
Gunicorn
psycopg2-binary
razorpay
google-genai
reportlab
itsdangerous
Pillow
Markdown
bleach
pygments
